Refactor: Rename pin_create to pin_edit
